[REFACTOR] Fix missing imports
