+++ src/client/react/AdminRoot.js
@@ -3,8 +3,8 @@ import { FocusStyleManager } from '@blueprintjs/core';
 import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
 import Loading from './pages/Loading';
-import LoginPage from './admin/LoginPage';
-import DashboardPage from './admin/DashboardPage';
+import LoginPage from './admin/pages/LoginPage';
+import DashboardPage from './admin/pages/DashboardPage';
 import NotFoundPage from './pages/NotFound';
 import { loadState } from '../actions/stateActions';
 import AppContainer from '../../../lib/react/components/AppContainer';
